requirements.txt
gradio
openai
vllm
fschat
auto-gptq
optimum
gradio_client 
pyyaml
charset-normalizer~=2.1.0
chardet